+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.6
 
+- Added: selected? and deselected? assertions for checkboxes.
 - Fixed a bug in each= so it respects the assert function passed as an argument (previously it was ignored and text= was always used).
 
 ## 0.1.5